pyproject.toml
[project]
name = "mini-rag-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "groq>=1.0.0",
    "numpy>=2.4.2",
    "pypdf>=6.7.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.54.0",
]
